+++ app/Models/Despacho/Registrador.php
@@ -2,12 +2,15 @@
 
 namespace App\Models\Despacho;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 
 class Registrador extends Model
 {
+    use HasFactory;
+
     protected $connection = 'despacho';
     protected $table = 'registradores';